requirements.txt
# Core dependencies
py-clob-client>=0.18.0    # Polymarket CLOB client
python-dotenv>=1.0.0       # Load .env files
pyyaml>=6.0                # YAML config parsing
requests>=2.31.0           # HTTP requests for BTC price feeds
websockets>=12.0           # WebSocket connections

# Dashboard
flask>=3.0.0               # Web dashboard framework
flask-login>=0.6.3         # Session management

# Data & utilities
sqlite3worker>=1.1.7       # Thread-safe SQLite (stdlib sqlite3 used directly)
schedule>=1.2.0            # Periodic task scheduling

# Crypto
eth-account>=0.11.0        # Ethereum account / signing
web3>=6.15.0               # Web3 for blockchain interactions
